refactor(client): widget-layer foundation for the list-shaped settings tabs
Preparation for migrating Ssh / Keybindings / Profiles onto the widget
layer. All three share one shape — a list, an edit panel for the selected
entry, Add/Delete buttons, and a delete dialog — and none of it is
expressible today.

Split `draw.rs` by control family. It sat at 779 lines against the
project's 800-line guideline, so the new control kinds would not have fit.
`draw/mod.rs` keeps the `draw_widget` dispatch and the chrome every control
shares (row fill, label column, focus ring, text baseline); `controls.rs`,
`slider.rs`, `text.rs` and `list.rs` own one family each. No behaviour
change: the moved code is byte-identical apart from visibility.

Widen `WidgetId.index` from u8 to u16. The migrated tabs are fixed-length
(14 rows at most), but a list-shaped category addresses one widget per
list entry and those lists are user-populated with no upper bound — a u8
would have stopped addressing entries past the 256th with no diagnostic.
`as_u32` shifts by 16 instead of 8; the widest offset is 0xFF_FFFF, well
inside the 100M-wide node-id slot reserved for SETTINGS_WIDGET_BASE. The
five per-tab focus counters that hold widget indices follow. Ssh's and
Keybindings' counters keep their own numbering until those tabs migrate.

Add the three missing `WidgetKind` arms — `ListItem`, `Button` and
`KeyCapture` — with their drawing and their AccessKit roles
(ListBoxOption / Button / TextInput). A key capture is not text entry: it
swallows the next key press rather than appending it, and a reader has to
announce that difference, so it does not reuse `Text`.

Give `Text` a real caret. The renderer appended `_` to the value and
ignored `TextInputState.cursor` entirely, so the caret sat at the end of
the string however the user had moved it — this already affected the
migrated Security byte-cap fields, the Startup shell fields and the font
fallback list. `caret` is a byte offset that falls back to end-of-string
when absent or mid-character.

Tests: 1651 pass (+15). IME preedit is still not shown in these fields;
that is pre-existing and left for the tab migrations.

via [Happy](https://happy.engineering)

+++ nexterm-client-gpu/src/renderer/event_handler/settings_panel_hit.rs
@@ -27,30 +27,30 @@ pub(super) enum SettingsPanelHit {
     /// UI/UX v3 P1c: click on a row inside the Font category. The payload is
     /// the widget index; a press on the size slider's track returns `Slider`
     /// instead so a drag can start.
-    FontRow(u8),
+    FontRow(u16),
     /// UI/UX v3 P1c: click on a row inside the Startup category.
-    StartupRow(u8),
+    StartupRow(u16),
     /// Theme color dot.
     ThemeColor(usize),
     /// UI/UX v3 P1b: click on a non-swatch row of the Theme category. The
     /// payload is the widget index (`THEME_SCHEME` / `THEME_FOLLOW_SYSTEM`);
     /// clicking focuses the row and, for the toggle, flips it.
-    ThemeRow(u8),
+    ThemeRow(u16),
     /// Click on a row inside the Window category. The payload is the widget
     /// index; see `widgets::settings_window::row` for the names. Clicking a
     /// toggle or cycler row also changes its value, while numeric rows only
     /// take focus (a press on a slider track returns `Slider` instead, so a
     /// drag can start).
-    WindowRow(u8),
+    WindowRow(u16),
     /// Phase 2c follow-up: click on a row inside the Blocks category.
     /// row 0 = `blocks_enabled` toggle, row 1 = increment `border_width_px`
     /// (wraps from 8 back to 1), row 2 = `show_exit_code_badge` toggle.
     /// The "Editing" hint row (3) intentionally has no hit zone.
-    BlocksRow(u8),
+    BlocksRow(u16),
     /// Click on a row inside the Security category. row 0..=3 are consent-policy
     /// cyclers (click cycles forward), 4..=6 are byte-cap fields (click focuses
     /// and starts editing). The footer note row has no hit zone.
-    SecurityRow(u8),
+    SecurityRow(u16),
     /// P4 (WT-like UX): the "Open config.toml" link in the footer bar.
     OpenConfigFile,
     /// P2-A (WT-like UX): the "reset category to defaults" link in the
